+++ ui/package.json
@@ -79,6 +79,7 @@
     "enzyme": "^3.7.0",
     "enzyme-adapter-react-16": "^1.7.0",
     "highcharts": "^6.2.0",
+    "highcharts-react-official": "^2.0.0",
     "immutable": "^4.0.0-rc.9",
     "lodash": "^4.17.11",
     "moment": "^2.20.1",
