+++ ui/src/app/cohort-review/individual-participants-charts/individual-participants-charts.html
@@ -1,1 +1,2 @@
-<chart #chartRef [options]="chartOptions" (load)="getChartObj($event.context)" style="display:block; width: 90% !important;" type="chart"></chart>
+<!--<chart #chartRef [options]="chartOptions" (load)="getChartObj($event.context)" style="display:block; width: 90% !important;" type="chart"></chart>-->
+<div id="react-chart" ></div>
